+++ src/macros.rs
@@ -1,0 +1,57 @@
+#![macro_use]
+
+// Loops through the attributes once and pulls out the ones we ask it to. It
+// will check that the required ones are there. This could have been done with
+// attrs.find but that would be inefficient.
+//
+// This is probably a really terrible way to do this. It does cut down on lines
+// though which is nice.
+macro_rules! get_attrs {
+    ($attrs:expr, optionals: [$(($oName:pat, $oVar:ident, $oMethod:expr)),*],
+     required: [$(($name:pat, $var:ident, $method:expr)),*], $err:expr) => {
+        {
+            $(let mut $oVar = None;)*
+            $(let mut $var = None;)*
+            for attr in $attrs.iter() {
+                match attr.name.local_name.as_ref() {
+                    $($oName => $oVar = $oMethod(attr.value.clone()),)*
+                    $($name => $var = $method(attr.value.clone()),)*
+                    _ => {}
+                }
+            }
+            if !(true $(&& $var.is_some())*) {
+                return Err($err);
+            }
+            (($($oVar),*), ($($var.unwrap()),*))
+        }
+    }
+}
+
+// Goes through the children of the tag and will call the correct function for
+// that child. Closes the tag
+//
+// Not quite as bad.
+macro_rules! parse_tag {
+    ($parser:expr, $close_tag:expr, $($open_tag:expr => $open_method:expr),*) => {
+        loop {
+            match try!($parser.next().map_err(TiledError::XmlDecodingError)) {
+                XmlEvent::StartElement {name, attributes, ..} => {
+                    if false {}
+                    $(else if name.local_name == $open_tag {
+                        match $open_method(attributes) {
+                            Ok(()) => {},
+                            Err(e) => return Err(e)
+                        };
+                    })*
+                }
+                XmlEvent::EndElement {name, ..} => {
+                    if name.local_name == $close_tag {
+                        break;
+                    }
+                }
+                XmlEvent::EndDocument => return Err(TiledError::PrematureEnd("Document ended before we expected.".to_string())),
+                _ => {}
+            }
+        }
+    }
+}
