+++ tests/lib.rs
@@ -2,7 +2,7 @@ extern crate tiled;
 
 use std::path::Path;
 use std::fs::File;
-use tiled::{Map, TiledError, PropertyValue, parse, parse_file, parse_tileset};
+use tiled::{parse, parse_file, parse_tileset, Map, PropertyValue, TiledError};
 
 fn read_from_file(p: &Path) -> Result<Map, TiledError> {
     let file = File::open(p).unwrap();
@@ -45,23 +45,31 @@ fn test_image_layers() {
     {
         let first = &r.image_layers[0];
         assert_eq!(first.name, "Image Layer 1");
-        assert!(first.image.is_none(), "{}'s image should be None", first.name);
+        assert!(
+            first.image.is_none(),
+            "{}'s image should be None",
+            first.name
+        );
     }
     {
         let second = &r.image_layers[1];
         assert_eq!(second.name, "Image Layer 2");
-        let image = second.image.as_ref().expect(&format!("{}'s image shouldn't be None", second.name));
+        let image = second
+            .image
+            .as_ref()
+            .expect(&format!("{}'s image shouldn't be None", second.name));
         assert_eq!(image.source, "tilesheet.png");
         assert_eq!(image.width, 448);
         assert_eq!(image.height, 192);
     }
 }
 
-
 #[test]
 fn test_tile_property() {
     let r = read_from_file(&Path::new("assets/tiled_base64.tmx")).unwrap();
-    let prop_value: String = if let Some(&PropertyValue::StringValue(ref v)) = r.tilesets[0].tiles[0].properties.get("a tile property") {
+    let prop_value: String = if let Some(&PropertyValue::StringValue(ref v)) =
+        r.tilesets[0].tiles[0].properties.get("a tile property")
+    {
         v.clone()
     } else {
         String::new()
